feat: create and move location crate
