+++ crates/tinymist-analysis/Cargo.toml
@@ -19,6 +19,7 @@ serde.workspace = true
 strum.workspace = true
 toml.workspace = true
 typst.workspace = true
+log.workspace = true
 
 [dev-dependencies]
 insta.workspace = true
